Do not add PR creator as reviewer
Github API returns all the users who participate in a review process
when we use the /reviews endpoint. As a result, if the author of a PR
also comments inside the review comment of a reviewer (e.g. in code
comments), then our workflow will add the author of the PR as a reviewer
too.

For that purpose, we need to check if the author is the same as the
reviewer right before applying the reviewer-by git trailers.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>