Change value of items per page, when requesting reviews
Github API uses pagination and therefore when we have a lot of comments
in a review, we might not be able to get the final APPROVED review. For
that purpose, we change the number of items per page when we request the
reviews from the github API to 300. Hopefully, this value will be
enough.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>